+++ next.config.ts
@@ -1,5 +1,13 @@
+import withSerwistInit from "@serwist/next";
 import type { NextConfig } from "next";
+
+const withSerwist = withSerwistInit({
+  swSrc: "app/sw.ts",
+  swDest: "public/sw.js",
+  disable: process.env.NODE_ENV === "development",
+  additionalPrecacheEntries: [{ url: "/~offline" }]
+});
 
 const nextConfig: NextConfig = {};
 
-export default nextConfig;
+export default withSerwist(nextConfig);